Tests/LibWeb/Text/expected/HTML/Element-checkVisibility.txt

content-visibility-parent hidden: false
content-visibility-child hidden: true
